Refactor setting page
