+++ src/components/Settings/index.ts
@@ -1,0 +1,13 @@
+export { default as BlogSettingsForm } from "./BlogSettingsForm";
+export { default as AccountSettingsForm } from "./AccountSettingsForm";
+export { default as SecuritySettingsForm } from "./SecuritySettingsForm";
+export { default as SettingsTabNavigation } from "./SettingsTabNavigation";
+export { default as SettingsHeader } from "./SettingsHeader";
+export { default as ErrorAlert } from "./ErrorAlert";
+export { default as SocialLinksManager } from "./SocialLinksManager";
+
+// Export types
+export type { FormData as BlogFormData } from "./BlogSettingsForm";
+export type { UserProfile, UserFormData } from "./AccountSettingsForm";
+export type { PasswordFormData } from "./SecuritySettingsForm";
+export type { SettingsTab } from "./SettingsTabNavigation";
