+++ src/types/article.ts
@@ -1,40 +1,3 @@
-export type ApiArticle = {
-  articleId: string;
-  authorId: string;
-  title: string;
-  subtitle: string;
-  summary: string;
-  content: string;
-  slug: string;
-  thumbnailUrl: string;
-  coverImageUrl: string;
-  timeToReadInMinute: number;
-  likes: number;
-  tagIds: string[];
-  tagSlugs: string[];
-  status: ArticleStatus;
-  createdAt: string; // ISO string
-  updatedAt: string; // ISO string
-  publishedAt: string | null;
-  archivedAt: string | null;
-};
-
-export type ApiArticles = {
-  page: number;
-  pageSize: number;
-  totalCount: number;
-  results: ApiArticle[];
-};
-
-export type ArticleFilter = {
-  Keyword: string;
-  Statuses: string[];
-  TagIds: string[];
-  SortBy: string;
-  Page: number;
-  PageSize: number;
-};
-
 export type Article = {
   articleId: string;
   authorId: string;
@@ -56,6 +19,15 @@ export type Article = {
   archivedAt: string | null;
 };
 
+export type ArticleFilter = {
+  Keyword: string;
+  Statuses: string[];
+  TagIds: string[];
+  SortBy: string;
+  Page: number;
+  PageSize: number;
+};
+
 // FIX Modify to use PublishedAt for sorting
 export enum ArticleSortBy {
   Latest = "CreatedAtNewest",
@@ -67,14 +39,4 @@ export enum ArticleStatus {
   Draft = "Draft",
   Published = "Published",
   Archived = "Archived",
-}
-
-export interface CreateArticleApiRequest {
-  title: string;
-  subtitle?: string;
-  summary?: string;
-  content: string;
-  coverImageUrl?: string;
-  thumbnailUrl?: string;
-  tagIds?: string[];
 }
